Getting all cuddly with a monad
I am much closer to understanding them. Really a monad is just a bit of
boilerplate that places one thing in a structure and provides a method
for operating on the structure and value.

They find nice applications in exception handling, logging, IO, etc. But
they needn't be used in such ways. They could also be used to couple two
distinct operations. Or so I think.

Then I found that all the code RWH is out of date. The language recently
reclassed Monad as a subclass of Applicative. I'll go off and dig around
with Functors and Applicatives for awhile now ... then I'll come back
and mess with monads.