Add instance of show function for simple parser
Previously I had used the automatically generated show function (e.g.
`deriving(show)`). This approach had the useful debugging feature of
including the constructor in the printed output.

I have now replaced it with a show instance based on the one in the
tutorial. At this point, it is not an improvement, but I will be
building on it.