Implemented WYAS through ch3a, but am lost
OK, there is a growingly obvious dearth in my Haskell knowledge. I can
copy this stuff over, I can even understand it, but I couldn't write it.
I need to start just doing things, implemenenting on my own, then come
back and do this tutorial.

It is suppose to be for someone who knows nothing about Haskell. Such a
person might be able to copy the code and make it run, but that wouldn't
make them capably of writing their own compiler.

So, I'll pop over to compiler-practice repo and play there.