Add Data.Graph experiments
I am not sure I like the way it works. The simple `graphFromEdges`
functions returns three things: the graph, a function for mapping keys
to `Maybe Vertex`, and a function for mapping a Vertex to a `(node, key,
[key])` tuples. It feels a little awkward.

There is a Data.Graph.Algorithm module which may alleviate some of the
difficulty, but I don't have it installed and currently have no internet
access (wifi from the backseat of a van in a Chinese alley is pretty
crumby).