Add tree with Monoid type
Monoids are systems that have an identity element and an associative
operator that is closed and conserves type.

I studied this before and it seems terribly esoteric. But considering
whether a system is a monoid, and how and whether to make it one, is a
deep pursuit. A graph is an example of this.

Adding an identity element to a graph is not hard (it is just an empty
graph), but is something that one might never think to do otherwise.

But adding an associative operator that yields a new graph is a deep
graph theory problem. There are several reaonsable operators. I
implemented a simple system where `mappend` adds together the input
nodes (appending their values and taking the union of their children, in
order). Since I append the values, the values must be monoids.

There are certainly other ways to add two graphs. For example, taking
the union of their edges. This would be a different monoid.