Add functions to tree
Temporarily gave up on making the tree an instance of Applicative or
Monad. I wrote two very useful functions using simple recursion.

Since the tree is an instance of functor, I can use fmap:

 (a -> b) -> f a -> f b

But sometimes I want to condition on the container, for example

 (a -> [a] -> b) -> Tree a -> Tree b

Where the [a] is the list of the values of the children.

Of I might want to condition on a different permutation:

 ([a] -> b) -> Tree a -> Tree b

In this case, each element in the functor as a single value (a), but the
container has one relevant value also, the children. So there are three
permutations.

```
 (a -> b)        -> f a -> f b  -- operate on the value (fmap)
 ([a] -> b)      -> f a -> f b  -- operate on the children
 (a -> [a] -> b) -> f a -> f b  -- join operation on kids and parents
```

The latter two operations do preserve the structure on which they
operate (as required by a functor).

I'm not sure how to generalize this.

Changing subject, I would like to write the function

```
all (\= "jes") tree
```

However this requires that Tree be an instance of Foldable. I can do
this now with

```
all $ tree2list $ fmap (\= "jes") tree
```

The fact that I can do this so easily, probably means I could easily
make Tree an instance of foldable. I'll look into this next.