Wrote a tree zip function
I feel like this form can be generalized into a generic functional form.
Though I am not sure which. Basically it is a zip operation where the
first zippant is a series, defined by a successor function.

```
    zip :: (a -> a) -> a -> Tree b -> Tree (a,b)
```

This function could be useful, for example, if you wanted to attach a
unique id to every element in a tree.