Added basic file IO using System.Process
A really cool thing about Haskell is that there are formal methods for
deciding what works, and all that is needed is the type signatures. But
I need to work out a slightly better syntax for my type proofs. I am
roughly following the methods I've gleaned from a few type theory books,
but need to review.