Use ExisitentialQuantification to allow Hetero log
I wan to be able to store every intermediate value. However, I don't
want to limit the pipeline to a single type, so I need a heterolist.
This is made possible with existential types.

Adding new functions to pipelines though, is getting really heavy. For
example, just coupling a value to the current value, requires this
monstrosity:

```
couple' :: (Show a, Show b) => a -> b -> Thread Log (a,b)
couple' x y = Thread (Right (log, (x,y))) where
  log = Log [("Couple with " ++ show x, Just (Het (x,y)))]
```

There is a lot of boilerplate here that I should be able to remove. I'll
do that next.