Add toy code for serialization and md5
This is was my first time to `cabal init`. It generates nice comments in
the cabal file, though I may disable them in the future in favor of the
minimal form. It also generates a `ChangeLog.md` file, which I don't
think I use (that's what git commit messages are for, if I ever need to
make a change log, I can just dump them into a file). I also don't need
a `Setup.hs` file, maybe later (not sure what it is even for).

And another meta comment: I think for commits, especially in practice
repos, I should use hashtags. This will allow me to more easily search
my github history. For this to really be maximally useful, I should have
a defined set of tags, but that is a problem for later.

  #serialization #DeriveGeneric #Data.Binary #md5 #digest